refactor: organize project files
