+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:49881/","mongoDBName":"jest"}
+{"mongoUri":"mongodb://127.0.0.1:53485/","mongoDBName":"jest"}
