+++ src/data/protocols/db/survey-result/save-survey-result-repository.ts
@@ -1,5 +1,5 @@
 import { SurveyResultModel } from '@domain/models/survey-result'
-import { SaveSurveyResultModel } from '@domain/usecases/save-survey-result'
+import { SaveSurveyResultModel } from '@domain/usecases/survey-result/save-survey-result'
 
 export interface SaveSurveyResultRepository {
   save(data: SaveSurveyResultModel): Promise<SurveyResultModel>
