+++ src/data/protocols/db/survey/add-survey-repository.ts
@@ -1,4 +1,4 @@
-import { AddSurveyModel } from '@domain/usecases/add-survey'
+import { AddSurveyModel } from '@domain/usecases/survey/add-survey'
 
 export interface AddSurveyRepository {
   add(surveyData: AddSurveyModel): Promise<void>
