+++ src/data/usecases/account/authentication/db-authentication-protocols.ts
@@ -2,5 +2,5 @@ export * from '@data/protocols/cryptography/hash-comparer'
 export * from '@data/protocols/cryptography/encrypter'
 export * from '@data/protocols/db/account/load-account-by-email-repository'
 export * from '@data/protocols/db/account/update-access-token-repository'
-export * from '@domain/usecases/authentication'
+export * from '@domain/usecases/account/authentication'
 export * from '../add-account/db-add-account-protocols'
