+++ src/data/usecases/account/load-account-by-token/db-load-account-by-token-protocols.ts
