+++ src/data/usecases/account/load-account-by-token/db-load-account-by-token.ts
@@ -1,4 +1,4 @@
-import { LoadAccountByToken } from '@domain/usecases/load-account-by-token'
+import { LoadAccountByToken } from '@domain/usecases/account/load-account-by-token'
 import { Decrypter, LoadAccountByTokenRepository, AccountModel } from './db-load-account-by-token-protocols'
 
 export class DbLoadAccountByToken implements LoadAccountByToken {
