--- src/data/usecases/save-survey-result/db-save-survey-result-protocols.ts
@@ -1,4 +1,0 @@
-export * from '@domain/models/survey-result'
-export * from '@domain/usecases/save-survey-result'
-export * from '@data/protocols/db/survey/save-survey-result-repository'
-export * from './db-save-survey-result'
