+++ src/data/usecases/survey/add-survey/db-add-survey-protocols.ts
@@ -1,2 +1,2 @@
-export * from '@domain/usecases/add-survey'
+export * from '@domain/usecases/survey/add-survey'
 export * from '@data/protocols/db/survey/add-survey-repository'
