+++ src/data/usecases/survey/load-survey-by-id/db-load-survey-by-id.ts
@@ -1,4 +1,4 @@
-import { LoadSurveyById } from '@domain/usecases/load-survey-by-id'
+import { LoadSurveyById } from '@domain/usecases/survey/load-survey-by-id'
 import { LoadSurveyByIdRepository, SurveyModel } from './db-load-survey-by-id-protocols'
 
 export class DbLoadSurveyById implements LoadSurveyById {
