+++ src/data/usecases/survey/load-surveys/db-load-surveys-protocols.ts
@@ -1,4 +1,4 @@
 export * from '@data/protocols/db/survey/load-survey-repository'
 export * from '@domain/models/survey'
-export * from '@domain/usecases/load-surveys'
+export * from '@domain/usecases/survey/load-surveys'
 export * from './db-load-surveys'
