+++ src/infrastructure/db/mongodb/account/account-mongo-repository.ts
@@ -3,7 +3,7 @@ import { LoadAccountByEmailRepository } from '@data/protocols/db/account/load-ac
 import { LoadAccountByTokenRepository } from '@data/protocols/db/account/load-account-by-token-repository copy'
 import { UpdateAccessTokenRepository } from '@data/protocols/db/account/update-access-token-repository'
 import { AccountModel } from '@domain/models/account'
-import { AddAccountModel } from '@domain/usecases/add-account'
+import { AddAccountModel } from '@domain/usecases/account/add-account'
 import { ObjectId } from 'mongodb'
 import { MongoHelper } from '../helpers/mongo-helper'
