+++ src/infrastructure/db/mongodb/survey-result/survey-result-mongo-repository.spec.ts
@@ -1,6 +1,6 @@
-import { AddAccountModel } from '@domain/usecases/add-account'
-import { AddSurveyModel } from '@domain/usecases/add-survey'
-import { SaveSurveyResultModel } from '@domain/usecases/save-survey-result'
+import { AddAccountModel } from '@domain/usecases/account/add-account'
+import { AddSurveyModel } from '@domain/usecases/survey/add-survey'
+import { SaveSurveyResultModel } from '@domain/usecases/survey-result/save-survey-result'
 import { Collection, ObjectId } from 'mongodb'
 import { MongoHelper } from '../helpers/mongo-helper'
 import { SurveyResultMongoRepository } from './survey-result-mongo-repository'
