+++ src/infrastructure/db/mongodb/survey-result/survey-result-mongo-repository.ts
@@ -2,7 +2,7 @@ import {
   SaveSurveyResultModel,
   SaveSurveyResultRepository,
   SurveyResultModel
-} from '@data/usecases/save-survey-result/db-save-survey-result-protocols'
+} from '@data/usecases/survey-result/save-survey-result/db-save-survey-result-protocols'
 import { MongoHelper } from '../helpers/mongo-helper'
 
 export class SurveyResultMongoRepository implements SaveSurveyResultRepository {
