+++ src/main/factories/usecases/account/add-account/db-add-account-factory.ts
@@ -1,5 +1,5 @@
-import { DbAddAccount } from '@data/usecases/add-account/db-add-account'
-import { AddAccount } from '@domain/usecases/add-account'
+import { DbAddAccount } from '@data/usecases/account/add-account/db-add-account'
+import { AddAccount } from '@domain/usecases/account/add-account'
 import { BcryptAdapter } from '@infrastructure/cryptography/bcrypt-adapter/bcrypt-adapter'
 import { AccountMongoRepository } from '@infrastructure/db/mongodb/account/account-mongo-repository'
