+++ src/main/factories/usecases/account/load-account-by-token/db-load-account-by-token-factory.ts
@@ -1,5 +1,5 @@
-import { DbLoadAccountByToken } from '@data/usecases/load-account-by-token/db-load-account-by-token'
-import { LoadAccountByToken } from '@domain/usecases/load-account-by-token'
+import { DbLoadAccountByToken } from '@data/usecases/account/load-account-by-token/db-load-account-by-token'
+import { LoadAccountByToken } from '@domain/usecases/account/load-account-by-token'
 import { JwtAdapter } from '@infrastructure/cryptography/jwt-adapter/jwt-adapter'
 import { AccountMongoRepository } from '@infrastructure/db/mongodb/account/account-mongo-repository'
 import env from '@main/config/env'
