+++ src/main/factories/usecases/survey/add-survey/db-add-account-factory.ts
@@ -1,5 +1,5 @@
-import { DbAddSurvey } from '@data/usecases/add-survey/db-add-survey'
-import { AddSurvey } from '@domain/usecases/add-survey'
+import { DbAddSurvey } from '@data/usecases/survey/add-survey/db-add-survey'
+import { AddSurvey } from '@domain/usecases/survey/add-survey'
 import { SurveyMongoRepository } from '@infrastructure/db/mongodb/survey/survey-mongo-repository'
 
 export const makeDbAddSurvey = (): AddSurvey => {
