+++ src/main/factories/usecases/survey/load-surveys/db-load-surveys-factory.ts
@@ -1,5 +1,5 @@
-import { DbLoadSurveys } from '@data/usecases/load-surveys/db-load-surveys'
-import { LoadSurveys } from '@domain/usecases/load-surveys'
+import { DbLoadSurveys } from '@data/usecases/survey/load-surveys/db-load-surveys'
+import { LoadSurveys } from '@domain/usecases/survey/load-surveys'
 import { SurveyMongoRepository } from '@infrastructure/db/mongodb/survey/survey-mongo-repository'
 
 export const makeDbLoadSurveys = (): LoadSurveys => {
