+++ src/presentation/controller/auth/login/login-controller-protocols.ts
@@ -1,2 +1,2 @@
 export * from '@presentation/protocols' // import generic protocols for controllers
-export * from '@domain/usecases/authentication';
+export * from '@domain/usecases/account/authentication';
