+++ src/presentation/controller/auth/signup/signup-controller-protocols.ts
@@ -1,4 +1,4 @@
 export * from '@presentation/protocols' // import generic protocols for controllers
-export * from '@domain/usecases/add-account'
+export * from '@domain/usecases/account/add-account'
 export * from '@domain/models/account'
-export * from '@domain/usecases/authentication'
+export * from '@domain/usecases/account/authentication'
