+++ src/presentation/controller/survey/add-survey/add-survey-controller-protocols.ts
@@ -1,2 +1,2 @@
 export * from '@presentation/protocols'
-export * from '@domain/usecases/add-survey'
+export * from '@domain/usecases/survey/add-survey'
