+++ src/presentation/controller/survey/load-surveys/load-surveys-controller-protocols.ts
@@ -1,3 +1,3 @@
 export * from '@presentation/protocols'
-export * from '@domain/usecases/load-surveys'
+export * from '@domain/usecases/survey/load-surveys'
 export * from '@domain/models/survey'
