+++ src/presentation/middlewares/auth-middleware-protocols.ts
@@ -1,6 +1,6 @@
 export * from '../errors'
 export * from './auth-middleware'
 export * from '@presentation/helpers/http/http-helper'
-export * from '@domain/usecases/load-account-by-token'
+export * from '@domain/usecases/account/load-account-by-token'
 export * from '@domain/models/account'
 export * from '@presentation/protocols'
